+++ tests/testapp/admin.py
@@ -33,4 +33,5 @@ class ArticleAdmin(ContentEditor):
     inlines = [
         JSONPluginInline.create(models.Text),
         JSONPluginInline.create(models.Download),
+        JSONPluginInline.create(models.Marker),
     ]